refactor(explorer): group components related to aggregator choice, data and refresh interval in a `ControlPanel` component
