+++ mithril-explorer/src/components/ControlPanel/AggregatorSetter/AddAggregatorModal.js
