+++ mithril-explorer/src/components/ControlPanel/AggregatorSetter/index.js
