+++ mithril-explorer/src/components/ControlPanel/EpochSettings/index.js
