+++ mithril-explorer/src/components/ControlPanel/IntervalSetter/index.js
